Updated Readme with new feature of filters, checks switched to abstract class
