+++ README.md
@@ -77,9 +77,54 @@ $this->taskSplitTestsByTime(5)
     ->run();
 ```
 
-this command need run all tests with `Codeception\Task\TimeReporter` for collect execution time. If you want just split tests between group (and not execute its) you can use SplitTestsByGroups.
+this command need run all tests with `Codeception\Task\TimeReporter` for collect execution time. If you want just split tests between group (and not execute its) you can use SplitTestsByGroups.  
+**Please be aware**: This task will not consider any 'depends' annotation!
 
+### Filter
 
+You can use a custom filter to select the necessary tests.
+
+Two filters already included: DefaultFilter, GroupFilter
+
+**DefaultFilter** is used by default and is every time the first filter which will be used.
+**GroupFilter** _Can only be used by taskSplitTestsByGroups_, allows you to filter the loaded tests by the given groups. You have the possibility to declare groups which you want to include or exclude. If you declare foo and bar as included then only tests with this both group annotations will be matched. The same thing is happend when you add excluded groups. Should you combine the included and excluded group the only tests which have exactly the correct group annotations for the included items and none of the excluded items.
+
+You can add as many filters as you want. The FIFO principle applies. The next filter will only get the result of the filter before.
+####USAGE:
+
+For example, you want all tests which have in the doc comment the groups 'foo' AND 'bar' but not 'baz' then you can do it like this:
+
+```php 
+$filter = new GroupFilter();
+$filter
+    ->groupIncluded('foo')
+    ->groupIncluded('bar')
+    ->groupExcluded('baz');
+
+$this->taskSplitTestsByGroups(5)
+   ->testsFrom('tests')
+   ->groupsTo('tests/_data/paratest_')
+   ->addFilter($filter)
+   ->run();
+```
+
+Now create your own filter class:
+```php 
+<?php
+
+declare(strict_types=1);
+
+namespace ...;
+
+use Codeception\Task\Filter\DefaultFilter;
+
+class CustomFilter extends DefaultFilter {
+
+}
+```
+
+The TestFileSplitterTask.php pushes an array of SplFileInfo Objects to the filter.  
+The TestsSplitterTask.php pushes an array of SelfDescribing Objects to the filter.
 ### MergeXmlReports
 
 Mergex several XML reports:
